Backend/Backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
Brotli==1.1.0
catboost==1.2.8
cffi==1.17.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
distlib==0.3.8
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
graphviz==0.20.3
html5lib==1.1
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
sklearn-compat==0.1.3
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.2
virtualenv==20.26.6
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
xgboost==3.0.0
zopfli==0.2.3
